+++ addon/components/bootstrap/control-wrapper.js
@@ -4,5 +4,6 @@ import layout from '../../templates/components/bootstrap/control-wrapper';
 
 export default Component.extend(InputableMixin, {
   layout,
+  classNames: ['form-group'],
   classNameBindings: ['floatingLabel:form-label-group'],
 });