+++ addon/components/bootstrap/form/-controls.js
@@ -1,0 +1,7 @@
+import Component from '@ember/component';
+import layout from '../../../templates/components/bootstrap/form/-controls';
+
+export default Component.extend({
+  layout,
+  tagName: '',
+});