+++ addon/components/bootstrap/modal/footer/-accept-button.js
@@ -1,7 +1,0 @@
-import Component from '@ember/component';
-import layout from '../../../../templates/components/bootstrap/modal/footer/-accept-button';
-
-export default Component.extend({
-  tagName: '',
-  layout
-});