+++ tests/integration/components/bootstrap/simple-form-test.js
@@ -2,7 +2,7 @@ import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 import a11yAudit from 'ember-a11y-testing/test-support/audit';
 
-moduleForComponent('<%= dasherizedModuleName %>', 'Integration | Component | <%= dasherizedModuleName %>', {
+moduleForComponent('bootstrap/simple-form', 'Integration | Component | bootstrap button', {
   integration: true
 });
 
@@ -11,16 +11,16 @@ test('it renders', function(assert) {
   // Handle any actions with this.on('myAction', function(val) { ... });
   // Example if your component had an `action` that it required:
   // this.set('noop', () => {});
-  // this.render(hbs`{{<%= dasherizedModuleName %> action=noop}}`);
-  this.render(hbs`{{<%= dasherizedModuleName %>}}`);
+  // this.render(hbs`{{bootstrap/simple-form action=noop}}`);
+  this.render(hbs`{{bootstrap/simple-form}}`);
 
   assert.equal(this.$().text().trim(), '');
 
   // Template block usage:
   this.render(hbs`
-    {{#<%= dasherizedModuleName %>}}
+    {{#bootstrap/simple-form}}
       template block text
-    {{/<%= dasherizedModuleName %>}}
+    {{/bootstrap/simple-form}}
   `);
 
   assert.equal(this.$().text().trim(), 'template block text');